Add reservation for ixa that uses the demand register.
If a demand register is passed into the operator ixa, then it just uses it
without reserving it. This causes issues because of the right side allocation
that follows. The answer was to use resreg() to reserve the register in this
case.

There are several other operands that need this treatment. These need testing.
It will follow.