Remove memory limits on pgen; define new flag to override self compile size.
pgen does not need a memory limit, because it doesn't store anything. I set it
to maxint, which on a 64 bit machine is likely more than the addressable store.
Will be necessary to revisit on a 32 bit or lower machine.

In pint, the store was shorted for self compile mode, which had more than one
effect. Now conditioned with BIG_STORE, which overrides that for storage.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Changed maxstr to be alias of maxint.

	modified:   source/pint.pas

1. Added BIG_STORE flag to override self_compile storage effect.