Fix headers for p2; flag error on missing filename in header.
 Changes to be committed:
	modified:   p2/pcomp.pas

1. Restore header configuration to ISO 7185. This will cause issues with gpc
host branch, I don't have an answer at the moment, see below.

	modified:   pascaline_tests/headertest2
	modified:   regress_report.txt
	modified:   source/pint.pas

1. Flag an error on external header filename assigns that come up empty.

Notes:

P2 doesn't have a macro processor. I restored it to ISO 7185 mode, but that will
break the gpc host branch. If that is an issue, we can put the macro processor
on P2.

This issue will also exist with P4.