Change the main entry point to external file; update scripts and makefile for that; regress.
A required step to modularity is the stacking system described elsewhere. The
location of "main" in the code is changed to an external assembly file, and that
is placed at the start of the module stack.

Accomodating this change with the current makefile structure was difficult, so
I moved to refactor the makefile to use .o files to form modules and use those
in the build sections of the makefile. Thus the makefile gets a needed overhaul.

Note that this change is not backward compatible.

 Changes to be committed:
	modified:   bin/compile
	modified:   bin/p6
	modified:   bin/testp2
	modified:   bin/testp4
	modified:   bin/testp5
	modified:   bin/testp6
	modified:   hosts/iso7185/compile
	modified:   hosts/iso7185/linux_X86/genobj
	modified:   hosts/iso7185/linux_X86/pcom
	modified:   hosts/iso7185/linux_X86/pgen
	modified:   hosts/iso7185/linux_X86/pint
	modified:   hosts/iso7185/linux_X86/pmach
	modified:   hosts/iso7185/linux_X86/spew
	modified:   hosts/iso7185/testp2
	modified:   hosts/iso7185/testp4
	modified:   hosts/iso7185/testp5
	modified:   hosts/iso7185/testp6

1. Update script files to use new main layout.

	modified:   hosts/iso7185/bit64/Makefile

1. Extensively refactored the makefile to create .o products and use them in the
builds instead of recompiling each time. Need to watch customization options
that end up in the product files.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/AMD64/gcc/pgen.pas

1. Remove "main" declaration and instead use the module name.