Update binaries; regress; allow eof to terminate intermediate file.
pcom does not generate the 'q' or quit instruction on modules. The reason for
that is the intermediates are designed to be concatenated with the program
module at the end, and the q instruction really means go.

We can emulate this in pgen by simply letting eof end the file. Its questionable
design by that Scott guy, but we can let it go for now.