Establish PRT result for pgen; update binaries; Regress; clear result in sfr.
Established a result compare for PRT in pgen so that the regression is stable.
Because pgen does not check for undefineds, test PRT 1918 was giving random
results in pgen. This is because the function result was left undefined. This
was fixed by clearing the result in SFR. Its tied to the debug checks option.

 Changes to be committed:
	modified:   bin/runprt

1. Check against different result for pgen mode.

	modified:   hosts/iso7185/linux_X86/genobj
	modified:   hosts/iso7185/linux_X86/pcom
	modified:   hosts/iso7185/linux_X86/pgen
	modified:   hosts/iso7185/linux_X86/pint
	modified:   hosts/iso7185/linux_X86/pmach

1. Update binaries.

	modified:   regress_report.txt

1. Run regression.

	modified:   source/AMD64/gcc/pgen.pas

1. In sfr, clear the function result. This is tied to the dbgchk option.