Update binaries; regress; add mpc instruction to pgen, pcom and pint; fix ujp; fix suv.
There are a few places where fat pointers, address+length, are constructed from an
address and a length or two addresses. On the stack machines this is identical
to a fat pointer, but pgen expects it to be a double register entry.

We introduce a new instruction, mpc or "make pointer from components" that takes
the two stack top components and makes a double register result. On the
interpreters, this has no function and is treated as a no-op.

 Changes to be committed:
	modified:   hosts/iso7185/linux_X86/genobj
	modified:   hosts/iso7185/linux_X86/pcom
	modified:   hosts/iso7185/linux_X86/pgen
	modified:   hosts/iso7185/linux_X86/pint
	modified:   hosts/iso7185/linux_X86/pmach

1. Update binaries.

	modified:   regress_report.txt

1. Ran regression.

	modified:   source/AMD64/gcc/pgen.pas

1. Add mpc instruction.

2. Remove botstk() bottom of stack check from ujp, turns out it does that with
things on the stack.

3. Fix suv overread.

	modified:   source/pcom.pas

1. Add mpc instruction.

2. Generate mpc in containerop().

3. Generate mpc in selector().

	modified:   source/pint.pas

1. Add mpc instruction and discard.