Updated binaries; regress; implement lsp instruction.
lsa is already a pgen no-op. We add lsp, which is the same as the ldp
instruction. lsa and ldp belong to a special class of instructions that
perform operations in the interpreter but have no meaning in pgen,
just as the mpc instruction performs an operation in pgen but not the
interpreter. This bridges the need for the interpreter to arrange its
stack, which have no analog in pgen.

 Changes to be committed:
	modified:   hosts/iso7185/linux_X86/pcom
	modified:   hosts/iso7185/linux_X86/pgen
	modified:   hosts/iso7185/linux_X86/pint

1. Update binaries.

	modified:   regress_report.txt

1. Run regression.

	modified:   source/AMD64/gcc/pgen.pas

1. put cvf and lsp into the instruction table.

2. Add cvf as alias to cuv.

3. Add lsp as no-op.

	modified:   source/pcom.pas

1. Add lsa to intmsgneol().

2. Change ldp to lsp in containerop().

3. Add lsp to instruction tables.

	modified:   source/pint.pas

1. Add lsp to instruction tables.

2. Generate lsp as ldp in assembler.