Partial commit; restore testpascaline to single compares; reorder reads on stack.
There was a mismatch between fat pointers and read ordering, which ends up with
length before variable ordering. This causes problems in pgen. The answer was to
reorder reads to be in standard fat pointer order, which is address/length in
address ascending.

This is a series of partial commits that break regression.

 Changes to be committed:
	modified:   bin/testpascaline

1. Return the compare and dif files to single names. This is possible because
the results are binned now.

	modified:   source/pcom.pas

1. In readprocedure(), after strings are parsed and their address loaded, if
the string is not a container, load a length and make it a complex pointer. This
regularizes the format on stack.

	modified:   source/pint.pas

1. Change the stack order for rds, rdsf and rdsp system calls so that the
string pointers are in standard fat pointer order of address/length ascending.