Partial commit; change exception handling to use local temps instead of stack.
This is partial commit because pmach and cmach were not also updated. Will
fix in next commit.

 Changes to be committed:
	modified:   regress_report.txt

1. Ran regression, only pint runs work complete.

	modified:   source/AMD64/gcc/pgen.pas

1. Added sev and rev instructions, store exception vector and restore exception
vector. This is implemented in pgen as store to and from stack.

	modified:   source/pcom.pas

1. Added sev and rev instructions. These instructions store to and restore from
local stack temps. See below.

	modified:   source/pint.pas

1. Added sev and rev instructions as aliases to local store/retrieve loda and str
instructions.

Normally exceptions are passed on the stack. This creates problems for pgen, since
it is not a stack based system. Creating a special sev and rev instruction makes
the operation abstract, even though it still uses a stack passed word.