Regress; change psystem_asm.asm to psystem.asm.
The original rationale for psystem_asm.asm was when it would have been psystem.s,
and that would have been prone to accidental erasure. However, as a .asm file,
it is protected. psystem.asm is a logical extension of the functions in psystem.c,
and they are C callable. Also, using .asm to mean "original, not machine created,
assembly files" is a widely accepted convention.