Replace quanta string system with pstrings in pgen.
This change replaces the string quanta system with pstrings in Pascaline.
This dramatically simplifies string processing in pgen and makes it more
efficient.

Here I use the "standalone" rule, which is that we use custom string routines
in pgen instead of string library routines, even if they are identical. The idea
is that only the compiler is relied on, and not the ability to include libraries
or externally compiled code. This simplifies bootstrapping or similar operations.
In practice the amount of string support needed is quite low.

 Changes to be committed:
	modified:   source/AMD64/gcc/pgen.pas

1. Remove strvsp definition.

2. Replace all appearances of strvsp with pstring.

3. Remove string quanta routines.

4. Add lenp(), find length of padded string.

5. Add strl(), make pstring from substring. This is used both alone and as a
building block for str().

6. Add str(), make pstring. This is a classic Pascaline function that makes a
dynamic copy of a string as a pstring. This is how a lot of string processing
takes place.

7. Added strp(), make pstring from padded string.

8. Added writeq(), write string with escaped quotes. This is used to write out
string constants.

9. Added cat(), concatenate strings to pstring result. Another fundamental
building block for strings.

10. Refactor all string operations in terms of pstrings.

11. Use direct writes for pstrings instead of the quanta writes.