Regress; update binaries; fix issue with inclass() routine.
 On branch master
 Changes to be committed:
	modified:   hosts/pascaline/linux_X86/genobj
	modified:   hosts/pascaline/linux_X86/pcom
	modified:   hosts/pascaline/linux_X86/pgen
	modified:   hosts/pascaline/linux_X86/pint
	modified:   hosts/pascaline/linux_X86/pmach

1. Update binaries.

	modified:   regress_report.txt

1. Regress.

	modified:   source/pcom.pas

1. In inclass(), if the entry given is a procedure or function and overloads
are present, returns the last overload matching the class requirement. This is
not correct, as it returns an entry from the group list that is not the
id entry, and thus causes issues.

inclass() is a poor routine, it really should just return a boolean, since it
just returns the original entry or nil if the there is not a match.