Regress; update binaries; Extensive refactor for #190 forwarded overloads don't work.
This is a pretty radical refactor of procdeclaration(). See #190 for more details.

 On branch master
 Changes to be committed:
	modified:   bin/testpascaline

1. Should be able to control list from compile line, didn't work.

	modified:   hosts/pascaline/linux_X86/genobj
	modified:   hosts/pascaline/linux_X86/pcom
	modified:   hosts/pascaline/linux_X86/pgen
	modified:   hosts/pascaline/linux_X86/pint
	modified:   hosts/pascaline/linux_X86/pmach

1. Update binaries.

	modified:   regress_report.txt

1, Ran regression.

	modified:   source/pcom.pas

1. Added or changed errors.

2. Left prtpar() diagnostic in, will remove.

3. Moved repeated parameter list error from parameterlist() to
procdeclaration().

4. Refactored chkovlpar().

5. Refactored fndovlgrp().

6. Removed prcsetflg().These actions are done in line now.

7. Prodeclaration() was extensively refactored. Basically all the prep work is
now done at the top, and we create both a new entry as well as fetch any
previous entry, and parse and hold on to the parameter list.

What follows is a series of checks for valid combinations, and then we decide
what parameter list to keep and what to throw away. Then we do the same for
the entry just created.

The big difference with the old code is we don't make decisions until we have
completely parsed the entire procedure or function header.

The code bears a lot of construction comments. I will remove them in the next
push.

	modified:   standard_tests/iso7185prt.cmp
	modified:   standard_tests/iso7185prt0053.ecp
	modified:   standard_tests/iso7185prt0306.ecp

1. Updated compare files.