Regress; update binaries; fix issue with procedure/function parameters.
If the parameter to an overload is proc/func on call and the head of the actual
expression is a function, the winnowing process will advance trying to match
a procedure/function parameter, then fail with error. This was changed to backing
up to the starting match list and then match the parameter as normal.