Regress; update binaries; fix joins issue; change pc.pas libraries to joins mode.
Fixed issue with joins where a joined module with uses declarations in it causes
uses of those modules in the main program to be elided. A joins flushes all
out, including any uses modules. The normal mode with both uses and joins is to
prevent double inclusion by filename matching. However that should be prevented
for joined modules.

The solution was to track the depth of joins modules and store that level in all
included entries. Thus the duplicate search is excluded for any file entry that
shows a non-zero joins level.

 On branch master
 Changes to be committed:
	modified:   hosts/pascaline/linux_X86/pcom
	modified:   hosts/pascaline/linux_X86/pgen

1. Update binaries.

	modified:   libs/source/parse.pas

1. Update documentation at top.

	modified:   pc/pc.pas

1. Move parse, services and scanner to joins mode and add qualidents.

	modified:   regress_report.txt

1. Regress.

	modified:   source/pcom.pas

1. Add joins level to file entries.

2. Add global joins nesting count.

3. Add joins level to openinput() procedure, and set file entries with that.

4. In usesjoins(), the scnnam() function rejects duplicates only if joins
nesting for that entry is zero.

5. In usesjoins(), track joins levels and pass to openinput().

6. Init joins level.