Refactor pc.
 On branch master
 Changes to be committed:
	modified:   bin/pc
	modified:   pc/pc.pas

1. Remove psystem from stdio, it does nothing there, since it is inserted by
dolink().

2. Refactor fndlnk() to separate the .o and .a files, then put the .a files
last. See below.

Notes:

We have to have a system to classify files as C/assembly or Pascaline due to
the different startup methods. We use:

.o - Pascaline
.a - C/assembly

In the future we may want .a files to have Pascaline formatting. If that occurs,
we would need a way to indicate the C or Pascaline status of files, most likely
testing if a .pas file exists.