Regress; update binary; Fix uses issue in pcom.
The uses duplicate suppression didn't work. Refactored.

 On branch master
 Changes to be committed:
	modified:   hosts/pascaline/linux_X86/pcom

1. Update binary.

	modified:   regress_report.txt

1. Regress.

	modified:   source/pcom.pas

1. Remove join field from filrec. This system didn't work.

2. Added use flag and uselist link.

3. Remove joinsnest.

4. Remove joinlvl parameter from openinput.

5. Add isuse flag parameter to openinput.

6. In openinput, if the file is a uses, do an insert to the uses list of the
nearest non-uses file block.

7. In usesjoins(), refactor uses search to go to the first non-uses block (main
or joins) and search the uses list there for duplicate uses levels.

Notes:

The system now reflects the way the system works. Each main or joins level has a
list of uses blocks, and that mirrors the fact that any uses opened in a
non-uses level will be placed into that block, then the block (a joins) will be
moved to the pile.

This has the odd side effect that uses defines attached to a joins level can be
reached with qualident, a fact that is interesting and will be covered in the
future.