Regress; update binary; Add test cases to Pascaline tests; fix uses/joins bug.
Added test cases to Pascaline to cover the previous bug with far accesses.
Also fixed issue with uses/joins that was discovered during implementation of
that.

The issue was that an externally parsed module with uses or joins is not
expected to produce code output and thus does not accurately support the
stack depth tracking system. This causes the stack to be out of balance
after parsing an external module, and this produces a compiler fault.

The fix was forcing the stack offsets back to zero after a module is
parsed.

 On branch master
 Changes to be committed:
	modified:   hosts/pascaline/linux_X86/pcom

1. Update binary.

	modified:   pascaline_tests/pascaline.cmp
	modified:   pascaline_tests/pascaline.pas
	modified:   pascaline_tests/pascaline1.pas

1. Add new tests for external module access.

	modified:   regress_report.txt

1. Regress.

	modified:   source/pcom.pas

1. Zero topnew at the end of body() parse.