Regress; update binary; refactor pc passthrough options according to #247
 On branch master
 Changes to be committed:
	modified:   bin/pc
	modified:   hosts/pascaline/pc

1. Update binary for pc.

	modified:   pc/pc.pas

1. Remove redundant passthough flags.

2. Add full set of passthrough options and set variables (tells if the option
was selected or not).

3. setflg() was overloaded and refactored to handle new options and the set
flags.

4. Added putflg() routine to doact(), handles inserting flags into command
lines. The passthough options are not inserted if not set by the user. This
keeps defaults from being overridden.

5. Refactored the existing option selects in instruction files. The list of
options here is incomplete and needs further consideration.

6. All passthroughs set off except list. The list option was being used to
override pcom's list normally on. This was true since the original Pascal-P
compiler but does not work anymore, since regular compiles are not expected
to list. This will be reverted later when pcom gets fixed.

	modified:   regress_report.txt

1. Regress.