Regress; update binaries; Add mark assembly lines option.
pgen outputs a map on how to translate source file lines to assembly language
instructions. gdb uses this to do source level debugging like breakpoint
placement and stepping.

A side effect of this is gdb stops understanding assembly file line numbers.
This can make it difficult to debug at the assembly level, and also causes
gdb to malfunction in cases.

Thus this is made optional, and can then be removed if we want the old, assembly
level behavior.

 On branch master
 Changes to be committed:
	modified:   hosts/pascaline/linux_X86/pcom
	modified:   hosts/pascaline/linux_X86/pgen
	modified:   hosts/pascaline/linux_X86/pint
	modified:   hosts/pascaline/linux_X86/pmach
	modified:   hosts/pascaline/pc

1. Update binaries.

	modified:   regress_report.txt

1. Regress.

	modified:   source/AMD64/gcc/pgen.pas

1. Change reference error supression to use refer().

2. Add domrklin option flag and set default to true.

3. Supress file headers and lindig flag on domrklin false.

	modified:   source/parcmd.pas

4. Add mrkasslin option.

	modified:   source/pcom.pas

1. Don't set mrkasslin default state.