Regress; update binary; refactor pgen.
This refactor is discussed in #264 pgen refactor goals.

It looks fairly extreme, but it is basically moving things around.
Most of the CPU independent code is moved to independent.pas (whth much more to
do). Two problems with that were the goto 99 to abort, and a forward call to
assemble(). A new procedure, abort() was added, which is virtual and pgen
overrides that. The assemble() forward is solved the same way, by a virtual
override pair.